Recreated Package Manager fix
