+++ com.mirror.steamworks.net/FizzySteamworks.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bd78bb22ae3a60443a865fb2d7dfdafa
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
