+++ com.mirror.steamworks.net/FizzySteamworks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e6e55fca1f4e4844593a3dd65a3ecb39
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
