+++ com.mirror.steamworks.net/IClient.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c65d0a35946fe34aa415488cd674bce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
