+++ com.mirror.steamworks.net/IServer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ed34f1f9e087d74094368efc7a84aad
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
