+++ com.mirror.steamworks.net/LICENSE.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 80411239d210c5c4fb875c2d24c06fe0
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
