+++ com.mirror.steamworks.net/LegacyClient.cs
