+++ com.mirror.steamworks.net/LegacyClient.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 862fc5e604ebaa0438662870d4a14e43
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
