+++ com.mirror.steamworks.net/LegacyCommon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 805bd7b4762d722499a9f229c905a8d9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
