+++ com.mirror.steamworks.net/LegacyServer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f6b4fdab451e8c647b7e26d8c4d934f9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
