+++ com.mirror.steamworks.net/NextClient.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a76b4f832739fae40ac8bebdc30ad63d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
