+++ com.mirror.steamworks.net/NextCommon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 79668d9d397a0fc4982afadecd11c869
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
