+++ com.mirror.steamworks.net/NextServer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 84bd1cd26eae15647a6f0a27cc0e2b69
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
