+++ com.mirror.steamworks.net/README.md.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 76396690c716d454c93044b15adaaffd
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
