+++ com.mirror.steamworks.net/package.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d89fc70cedcb5d64b903514772ec2220
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
